+++ Source/Lib/Common/Dimensions/States/AppDimensionState.Reducer.cs
@@ -21,5 +21,13 @@ public partial record AppDimensionState
 			// Recreate the current state so the UI gets a state changed event
 			return inState with {};
 		}
+
+		[ReducerMethod(typeof(NotifyUserAgentResizeAction))]
+		public static AppDimensionState ReduceNotifyUserAgentResizeAction(
+			AppDimensionState inState)
+		{
+			// Recreate the current state so the UI gets a state changed event
+			return inState with {};
+		}
 	}
 }