+++ Docs/TextEditor/changelog.md
@@ -5,7 +5,21 @@ All notable changes to the `Luthetus.TextEditor` repository will be documented i
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.6.0] - 2024-09-05
+## [2.7.0] - 2024-09-21
+<details>
+  <summary>Click to show changes</summary>
+
+	### Fixed
+    - User Agent resize events trigger remeasure
+    - Changing text editor's font-size triggers re-measure 
+    - After re-remeasuring, reload the virtualization result.
+    - BrowserResizeInterop
+	- InputTextEditorFontSize.razor changes
+	- ScrollbarSection.razor vertical reset point while dragging
+	- DISTANCE_TO_RESET_SCROLL_POSITION is 300px
+</details>
+
+## [2.6.0] - 2024-09-19
 <details>
   <summary>Click to show changes</summary>